+++ src/oci/resource_manager/models/config_source_record.py
@@ -21,6 +21,10 @@ class ConfigSourceRecord(object):
     #: This constant has a value of "GIT_CONFIG_SOURCE"
     CONFIG_SOURCE_RECORD_TYPE_GIT_CONFIG_SOURCE = "GIT_CONFIG_SOURCE"
 
+    #: A constant which can be used with the config_source_record_type property of a ConfigSourceRecord.
+    #: This constant has a value of "OBJECT_STORAGE_CONFIG_SOURCE"
+    CONFIG_SOURCE_RECORD_TYPE_OBJECT_STORAGE_CONFIG_SOURCE = "OBJECT_STORAGE_CONFIG_SOURCE"
+
     def __init__(self, **kwargs):
         """
         Initializes a new ConfigSourceRecord object with values from keyword arguments. This class has the following subclasses and if you are using this class as input
@@ -28,12 +32,13 @@ class ConfigSourceRecord(object):
 
         * :class:`~oci.resource_manager.models.GitConfigSourceRecord`
         * :class:`~oci.resource_manager.models.ZipUploadConfigSourceRecord`
+        * :class:`~oci.resource_manager.models.ObjectStorageConfigSourceRecord`
 
         The following keyword arguments are supported (corresponding to the getters/setters of this class):
 
         :param config_source_record_type:
             The value to assign to the config_source_record_type property of this ConfigSourceRecord.
-            Allowed values for this property are: "ZIP_UPLOAD", "GIT_CONFIG_SOURCE", 'UNKNOWN_ENUM_VALUE'.
+            Allowed values for this property are: "ZIP_UPLOAD", "GIT_CONFIG_SOURCE", "OBJECT_STORAGE_CONFIG_SOURCE", 'UNKNOWN_ENUM_VALUE'.
             Any unrecognized values returned by a service will be mapped to 'UNKNOWN_ENUM_VALUE'.
         :type config_source_record_type: str
 
@@ -61,6 +66,9 @@ class ConfigSourceRecord(object):
 
         if type == 'ZIP_UPLOAD':
             return 'ZipUploadConfigSourceRecord'
+
+        if type == 'OBJECT_STORAGE_CONFIG_SOURCE':
+            return 'ObjectStorageConfigSourceRecord'
         else:
             return 'ConfigSourceRecord'
 
@@ -70,7 +78,7 @@ class ConfigSourceRecord(object):
         **[Required]** Gets the config_source_record_type of this ConfigSourceRecord.
         The type of configuration source to use for the Terraform configuration.
 
-        Allowed values for this property are: "ZIP_UPLOAD", "GIT_CONFIG_SOURCE", 'UNKNOWN_ENUM_VALUE'.
+        Allowed values for this property are: "ZIP_UPLOAD", "GIT_CONFIG_SOURCE", "OBJECT_STORAGE_CONFIG_SOURCE", 'UNKNOWN_ENUM_VALUE'.
         Any unrecognized values returned by a service will be mapped to 'UNKNOWN_ENUM_VALUE'.
 
 
@@ -89,7 +97,7 @@ class ConfigSourceRecord(object):
         :param config_source_record_type: The config_source_record_type of this ConfigSourceRecord.
         :type: str
         """
-        allowed_values = ["ZIP_UPLOAD", "GIT_CONFIG_SOURCE"]
+        allowed_values = ["ZIP_UPLOAD", "GIT_CONFIG_SOURCE", "OBJECT_STORAGE_CONFIG_SOURCE"]
         if not value_allowed_none_or_none_sentinel(config_source_record_type, allowed_values):
             config_source_record_type = 'UNKNOWN_ENUM_VALUE'
         self._config_source_record_type = config_source_record_type