+++ CHANGELOG.rst
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 
 The format is based on `Keep a Changelog <http://keepachangelog.com/>`_.
 ====================
+2.38.3 - 2021-05-18
+====================
+
+Added
+-----
+* Support for spark-submit compatible options in the Data Flow service
+* Support for Object Storage as a configuration source in the Resource Manager service
+
+====================
 2.38.2 - 2021-05-11
 ====================